MEGA-commit... For several reasons. This is up-to-date for week 14, minus HW10.
One is that my computer broke a few weeks back; I kept using it despite a
broken GPU, but it got worse and worse (crashes, artifacts)... In the end,
I managed to lend a computer so I could get mine fixed. During this time,
I couldn't use git.

What's more, when I was to commit my work since then, just now, the index
was corrupt. I tried to re-create it, but had other broken objects... So
I cloned the old stuff from github, replaced it with my code, and here
we are.

The *third* reason is that, as can be seen, I split everything into
separate files, to speed up compilation time when working. Having to
wait 5+ seconds (on a fast computer) for every tiny change gets annoying.
With this system, I only need to compile the chapter I'm working on, except
for a release of the notes.